requirements.txt
amplpy>=0.13.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
networkx>=3.0
openpyxl>=3.1.0
tqdm>=4.65.0
click>=8.1.0
pytest>=7.3.0
pytest-cov>=4.0.0
psutil>=5.9.0
scikit-learn>=1.3.0
pyyaml>=6.0.0
statsmodels>=0.14.0